maint(common): update build coordinator for `resources/teamcity` changes
This modifies the build triggers so that changes in `resources/teamcity`
don't trigger builds on all platforms anymore. If a change happened in
`resources/teamcity/$platform` or `resources/teamcity/includes` it will
trigger a build for `$platform`.

Since bash doesn't support negative look-aheads we use `grep` with the
Perl regex option.

Part-of: #13943
Build-bot: skip