maint(mac): make make-km-dmg.sh less quiet
When build.sh was converted to a builder-style script, the -quiet flag
was then passed into make-km-dmg.sh, which meant that it no longer gave
valuable diagnostic output. This returns the script to default levels of
output (there is also a -verbose or -debug level we could use). The
variable QUIET_FLAG appears to have been used only in that one place.