maint(web): allow to specify paths as arguments
This adds command line arguments for the paths to the `s.keyman.com` and
`help.keyman.com` repos.

This deviates from sticking strictly with what's on TC so that we can
pass the desired arguments right from the start and don't have to add
another check in the TC build steps.