maint(developer): address code review comments
Don't try to install dependencies on macOS - build agents on Windows
and macOS are currently configured manually at this point. We don't
want to be attempting to install dependencies on macOS automatically -
homebrew is extremely fragile for CI as it always updates everything to
latest version (which of course breaks many things).

(from https://github.com/keymanapp/keyman/pull/14059#pullrequestreview-2873213671)