maint(web): add some more arguments for web release build script
Previously rsync used some TC variables. Now with the build step being
a script the TC variables are no longer accessible from the script, so
we add new arguments to the build script and then use these when calling
rsync. A similar changer was already made previously for the developer
release build.

Follow-up-of: #14049
Related: #14059
Part-of: #13399
Test-bot: skip