maint(developer,web): use braces for variables in powershell scripts
This changes variable references in the powershell scripts to use braces
(`${FOO}` instead of `$FOO`). This solves the problem where powershell
considers the colon in the rsync command as being part of a variable.
This should fix release builds for Web and Developer.

Also commented some unused variables.

Follow-up-of: #14167,#14168
Part-of: #13399
Test-bot: skip