maint(resources): fix triggering of build by Build-bot
Previously, if we had a line "Build-bot: release android" we passed
the arguments after the colon as a single argument, at least on my
machine. This change fixes things so that we split the string into multiple
parts separated by space.

Also updated a unit test to cover this scenario.

Fixes: #14315
Test-bot: skip