maint(common): fix use of `jq.inc.sh` in git hooks
`jq.inc.sh` gets also used in some scripts that don't source our
builder scripts, e.g. in git hooks. This means that `builder_is_windows`
is not defined. This change replaces the call of that function with
a slightly modified version of the previous implementation.

Follow-up-of: #14324
Build-bot: skip
Test-bot: skip