maint(windows): fix calls of renamed functions
A recent change (#14324) consolidated (and thus renamed) some functions
that were in several places. A PR that got created earlier but merged
later (#14267) added new functionality that still used the old function
names. This lead to release build failures on Windows and Developer.

This change fixes the names of the called functions.

Fixes: #14412
Follow-up-of: #14324
Test-bot: skip