refactor(web): refactor tracked-context state and usage pattern
Note that this will temporarily cause predictive-text to throw out older context states (that may be valid rewind targets) in favor of newer ones (that are known to be invalid rewind targets) due to the current context-state caching logic.  That'll be resolved later, in an upcoming PR - and preferably, this one shouldn't merge without that one.