refactor(web): refactor aligned tokenization-transition operations into separate method
I've really been looking forward to figuring this one out.  This refactor spins off the bulk of `attemptMatchContext` into a _class_ method on `ContextTokenization` - `.transitionTo` - which uses alignment data to build the new tokenization for incoming context based on the old tokenization in a clear manner.

By factorizing the method in this manner, we pave the way to allow this operation to be performed for different tokenization patterns based on either the initial or the final context (or both!) in the future - something epic-dict-breaker models will likely need to leverage.  It also makes the roles of the two halves of the code _far_ clearer than it was before, in my opinion.