feat(developer): New Project Parameters dialog in web/svelte
This is a prototype of the New Project Parameters dialog rewritten in
svelte, and hosted as a web page within the existing
TfrmNewProjectParameters form in Keyman Developer IDE (TIKE). The intent
of this prototype is to explore ways of incrementally moving UX out of
Delphi and into cross-platform web-based systems. Thus, this is not
complete (e.g. Targets is unlinked, no bcp 47 picker, issues with the
Browse-for-folder dialog), and the choice of UI framework
(Bootstrap/Sveltestrap) and even of Svelte itself is all up for
discussion.

The Svelte project is also currently running only in dev-mode, and would
need to be integrated into the existing Keyman Developer Server. This is
all well understood (there are adapters to link between Svelte and
Express, which is our backend http server in KMDevServer).

In my opinion, this looks like a viable way forward to incrementally
rewrite parts of Keyman Developer. I look forward to other peoples'
thoughts on this broad-brushstroke design and concept.

This also comes out of the discussion we had on #14657, where validation
messages were added to the New Project Parameters Delphi form, and for
implementation cost reasons, were put into a much expanded form,
resulting in a lot of wasted space on the form -- the web paradigm
should serve us much better for this kind of thing.

Build-bot: skip
Test-bot: skip
See-also: #14657