maint(resources): honor configure comment character in commit message
Git allows to specify a comment character different from the default
`#`. This is helpful when using markdown in commit messages, especially
when adding a user test section that has to start with `# User testing`
in order for the test bot to detect it.

This change modifies the `prepare-commit-msg` hook and adds a check for
a configured `core.commentString` or `core.commentChar`. If both are not
set it falls back to the default `#` character.

You can set a different comment character with
`git config core.commentChar //`.

Build-bot: skip
Test-bot: skip