change(windows): include map2pdb generation for relevant Delphi projects
Adds a call to map2pdb for Delphi projects, if map2pdb is an executable
on the path. This way, we get .pdb files we can use for debugging and
for performance profiling. We may be able to use these in future also
for Sentry traces and eliminate tds2dbg, but I will revisit that later.

If map2dbg is available, then the build and install actions will copy
the PDB to be alongside the corresponding executable, making most debug
actions relatively seamless for finding symbols.

Relates-to: #14787
Build-bot: skip release:developer,windows
Test-bot: skip