tests are running with user root in the Docker container
There is an issue that the Docker containers cannot modify the
files on KEYMAN_ROOT, since the user running build and tests on
Docker does not have the same gid and uid as the host user starting
the container.
The issue is solved by running the tests iwth user root in the
container.

Test-bot: skip
Build-bot: skip