refactor(web): implement simple search-space reuse
This changes SearchPath to construct new instances whenever the path is extended, treating SearchPath as an immutable portion of the search graph that may be referenced by path extensions - new instances of SearchPath.

This, in turn, removes the need to clone SearchPath instances when new input is received for an incoming token; the original path's represention will remain unchanged and may also be reused by a new instance extending the graph for the newly-received input.

Relates-to: #14445

Build-bot: skip build:web
Test-bot: skip